全部功能基本完成

+++ keepgym-manage-service/src/main/java/com/lhj/keepgym/manage/service/impl/CourseServiceImpl.java
@@ -3,13 +3,18 @@ package com.lhj.keepgym.manage.service.impl;
 import com.alibaba.dubbo.config.annotation.Service;
 import com.lhj.keepgym.bean.Coach;
 import com.lhj.keepgym.bean.Course;
+import com.lhj.keepgym.bean.OrderCourse;
 import com.lhj.keepgym.manage.mapper.CoachMapper;
 import com.lhj.keepgym.manage.mapper.CourseMapper;
+import com.lhj.keepgym.manage.mapper.OrderCourseMapper;
 import com.lhj.keepgym.service.CourseService;
 import org.springframework.beans.factory.annotation.Autowired;
+import tk.mybatis.mapper.entity.Example;
 
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.Date;
+import java.util.Iterator;
 import java.util.List;
 
 @Service
@@ -20,6 +25,9 @@ public class CourseServiceImpl implements CourseService {
 
     @Autowired
     private CoachMapper coachMapper;
+
+    @Autowired
+    private OrderCourseMapper orderCourseMapper;
 
     @Override
     public List<Course> findAllCourse() {
@@ -64,9 +72,23 @@ public class CourseServiceImpl implements CourseService {
 
     @Override
     public String deleteCourse(String courseId) {
-        int i = courseMapper.deleteByPrimaryKey(courseId);
-        if(i>0){
-            return "success";
+        OrderCourse orderCourse = new OrderCourse();
+        Example example = new Example(OrderCourse.class);
+        example.createCriteria().andEqualTo("courseId",courseId);
+        List<OrderCourse> orderCourses = orderCourseMapper.selectByExample(example);
+        if(orderCourses!=null){
+            Iterator<OrderCourse> iterator = orderCourses.iterator();
+            while (iterator.hasNext()){
+                orderCourse = iterator.next();
+                int i = orderCourseMapper.delete(orderCourse);
+                if(i<=0){
+                    return "fail";
+                }
+            }
+            int i = courseMapper.deleteByPrimaryKey(courseId);
+            if(i>0){
+                return "success";
+            }
         }
         return "fail";
     }
@@ -76,6 +98,7 @@ public class CourseServiceImpl implements CourseService {
         SimpleDateFormat simpleDateFormat = new SimpleDateFormat("yyyy-MM-dd hh:mm");
         course.setStartTime(simpleDateFormat.parse(course.getStrStartTime()));
         course.setEndTime(simpleDateFormat.parse(course.getStrEndTime()));
+        course.setCreateTime(new Date());
         int i = courseMapper.insertSelective(course);
         if(i>0){
             return "success";
